ipc_logging: add support for ipc_logging bootparam
Add CONFIG_IPC_LOGGING_BOOTPARAM option. This option adds a kernel
parameter 'qcom_ipc_logging.enabled', which allows IPC Logging to be
enabled at boot. If this option is selected, IPC Logging functionality
is disabled by default and can be enabled with
qcom_ipc_logging.enabled=1 on the kernel command line. The purpose of
this option is to allow a single kernel image to be distributed with
IPC Logging built in, but not necessarily enabled.

Change-Id: I344bd1b86cd96da9afe343ececd0d232be1592bf
Signed-off-by: Johnathan Cobden-Nolan <quic_jcobdenn@quicinc.com>