bus: mhi: core: Resume success on SYS_ERR
When a crash occurs while in suspended state M3/D3hot, the host
only discovers SYS_ERR pm-state change on resume. Handle this
state as valid transition for resume procedure, giving a chance
to the MHI core to manage the state (sbl transition, etc...).

Also it is possible that MHI on device can enter the SYS_ERROR state
and lead to an mhi_pm_resume timeout while it waits for M0 event
to come in. Add an extra check to detect SYS_ERROR after the M0
timeout occurs to avoid sending resume failure to the runtime or
system PM framework.

Change-Id: I97a3d5371ec81ea9002e4b40f134daad69dd0c5e
Signed-off-by: Bhaumik Bhatt <quic_bbhatt@quicinc.com>
Signed-off-by: Nitesh Gupta <quic_nitegupt@quicinc.com>