pcie: msm: Initialize client_id in pcie_i2c_ctrl_probe()
client_id field in pcie_i2c_ctrl_probe() function is
initialized to an invalid id.

"enum i2c_client_id client_id" is declared and there is
possibility of using this variable without initialising it.
As "enum i2c_client_id" is a global enum, the default value
of client_id is 0 which is equal to I2C_CLIENT_ID_NTN3 and
condition (client_id == I2C_CLIENT_ID_NTN3) can be satisfied
and that is not true when client_id is not assigned with a
proper value.

Change-Id: Ie5c238dc03d8fae51fbc32252696538ef4df42eb
Signed-off-by: Sushrut Shree Trivedi <quic_sushruts@quicinc.com>
Signed-off-by: Naga Rashmi Ayiluri <quic_nayiluri@quicinc.com>
Signed-off-by: Feng Chen <quic_fch@quicinc.com>