serial: msm_geni_serial: Configure SE_DMA_IF_EN register
Currently for non-console usecase, SE_DMA_IF_EN register is not
being configured and we are dependent for this register to be
configured while TZ/BOOT image UART FW loading. Due to this we
encountered an issue where this register was not configured during
FW loading leading to UART_ERROR_BUFFER_OVERRUN error.

Irrespective of SE_DMA_IF_EN reg is being configured while FW
loading or not, this register will be configurefd during probe
for non-console usecase.

Change-Id: Id98af7e22857da203eeba59eb1eb3adddd3e1812
Signed-off-by: Prasanna S <quic_prass@quicinc.com>