xfrm: Fix IPv6 RT info for IPv6 packets sent to IPv4 tunnel
This takes care of the case, when a raw IPv6 packet is being
sent into IPv4 IPsec tunnel, e.g. packet offload with tunnel mode
and mixed IP family.

The output method ip6_output requires the IPv6 routing information
in the skb->dst. In particular the rt6i_dev member.

For this particular case we will extend the dst entry
to contain the required information.

Change-Id: I03f46d3ad58666575e1d09485430e9644c23c92d
Signed-off-by: Ilia Lin <quic_ilial@quicinc.com>