adsprpc: Handle UAF scenario in put_args
Currently, the DSP updates header buffers with unused DMA handle fds.
In the put_args section, if any DMA handle FDs are present in the
header buffer, the corresponding map is freed. However, since the
header buffer is exposed to users in unsigned PD, users can update
invalid FDs. If this invalid FD matches with any FD that is already
in use, it could lead to a use-after-free (UAF) vulnerability.
As a solution,add DMA handle references for DMA FDs, and the map for
the FD will be freed only when a reference is found.

Acked-by: quic_anane <quic_anane@quicinc.com>
Signed-off-by: Ansa Ahmed <quic_ansa@quicinc.com>
(cherry picked from commit 9c0bf0f7e035403a097a8ac8ca48f3215cc00a67)
Change-Id: I5ab24df57eaaabacf8c4ea42e378b406cd39c3ac
Signed-off-by: QU TAO <quic_taoq@quicinc.com>