serial: msm_geni_serial: Prevent uart register access while device suspend
While runtime suspend is in progress and before it completes,if
set_mctrl or geni_se_dump_dbg_regs is invoked, there is a possibility
that it may access the UART registers without the clocks on.

To address this issue, added mutual exclusion in set_mctrl and
geni_se_dump_dbg_regs functions from accessing the geni register
while suspend is in progress.

Change-Id: I818aa074c6e44d6ebc755f1fba6b7a7b0c432dbf
Signed-off-by: Krishna Chaithanya Reddy G <quic_kgangapu@quicinc.com>
Signed-off-by: Panicker Harish <quic_pharish@quicinc.com>