virtio_fastrpc: Fix to avoid untrusted pointer dereferenc
Currently, the compat ioctl call distinguishes itself using a global
flag. If a user sends a compat ioctl call followed by a normal ioctl
call, it may result in using a user passed address as a kernel address
in the fastrpcdriver. To address this issue, consider localizing the
compat flag for the ioctl call.

Change-Id: I07263fae36a16ca497a40319bee3e50741452389
Signed-off-by: Anvesh Jain P <quic_ajainp@quicinc.com>