msm: adsprpc: Do not search the global map in the process-specific list
If a user makes the ioctl call for the fastrpc_internal_mmap with the
global map flag, fd, and va corresponding to some map already present
in the process-specific list, then this map present in the process-
specific list could be added to the global list. Because global maps
are also searched in the process-specific list. If a map gets removed
from the global list and another concurrent thread is using the same
map for a process-specific use case, it could lead to a use-after-free.
Avoid searching the global map in the process-specific list.

Acked-by: Anand Kulkarni <quic_anankulk@quicinc.com>
Change-Id: I38f537e38d5c83e05d8d45e8f35dc075c08943b9
Signed-off-by: Karthik Sanagavarapu <quic_kartsana@quicinc.com>