power: reset: qcom-reboot-reason: add reboot reason bootparam
This feature adds a command line parameter 'qcom-reboot-reason.reason'
which can be used to pass the reboot reason on bootup. This allows
the driver to access the reboot reason on targets where the bootloader
clears the reason after reading it.

Also add a reason 'admin-trigger' which can be used to indicate an
intentional reboot by a system/admin framework to differentiate from
the normal/cold boot reason.

Change-Id: Ib9dd5bb74b2eeb530da6d4461be7f4cae7cc5b4d
Signed-off-by: Johnathan Cobden-Nolan <quic_jcobdenn@quicinc.com>