msm: ep_pcie: Allow clients to query the link state during non-PCIe boot
In the case of a non-PCIe boot configuration, the EP-PCIe driver needs
to just probe successfully instead of failing the probe. This is mainly
to meet the GCC sync state requirements. So, the EP-PCIe driver simply
returns from the probe (a dummy probe) without doing any real
initialization in the case of a non-PCIe boot.

However, PCIe client drivers (like MHI) do not have a clue whether the
PCIe driver performed a full-fledged probe or a dummy probe. If there is
a way for the client driver to know the type of probe, then the client
driver can take appropriate actions, such as limited initialization
needed for a non-PCIe boot scenario.

Therefore, in the case of a non-PCIe boot scenario, when the PCIe driver
is performing a dummy probe, let's set the link state as INVALID STATE
and register the driver. This way, client drivers can query the link
state, and if they find the link state set to INVALID STATE, they will
know it is a dummy probe and take appropriate actions.

Change-Id: I3e776228a130f2c0011e81f2f5ba6f62e34c6151
Signed-off-by: Sumit Kumar <quic_sumk@quicinc.com>