msm: ep_pcie: handle mmio backup/restore in ep_pcie
Changes are to backup and restore MHI mmio region during
D3_cold and link enablement.

In scenarios where the device undergoes a reboot, MHI is set to the
SYSERR state. If there is a D3_hot to D3_Cold transition before BME
is set, the MHI MMIO space is not backed up properly. When the host
sets the device to D0 subsequently, the MHI driver restores the MMIO
space with all zeros. This causes the device to miss sending the
SYSERR MSI to the host.

 ----------------      ------------      -----      -------
|Reboot of device| -> |Reaches HLOS| -> |D3hot| -> |D3_cold|-
|post FW upgrade |     ------------      -----      -------  |
 ----------------                                            |
                                   -------------      ---    |
                                  | mmio restore| <- | D0|<- |
                                   -------------      ---

Change is to do backup/restore in ep_pcie entirely not depending
on MHI driver for the same. Implementation assumes that mmio
is retained in M2, L1ss sleep state.

 -------      -----------      ---------------      ------------
|D3_cold| -> |mmio backup| -> |Perst de-assert| -> |mmio restore|
 -------      -----------      ---------------      ------------ .

Change-Id: Iae5e7c891e63544f83a40f6b8bc9e6827308c929
Signed-off-by: Subramanian Ananthanarayanan <quic_skananth@quicinc.com>