serial: msm-geni-serial: Add atomic flag check for resource management
Warning messages were observed on suspend of the msm serial driver.
This patch addresses warning messages observed during the suspension
of the driver. It introduces a conditional check of the atomic flag to
manage resource enable/disable and SE clock operations. By updating the
state of the atomic flag, the issue is resolved, and the warning
messages are eliminated.

Change-Id: I0b506bb9ed32bbe1cfe7f1142cb9cdac8af20841
Signed-off-by: Zahir Shabbir Khan <quic_zshabbir@quicinc.com>