spmi: spmi-pmic-arb: Add support for interrupt recovery from hibernate
When SoC goes to hibernate, the individual PMIC devices under the
main SPMI PMIC arbiter may have their interrupts masked and when the
SoC is woken from hibernate state, the PMIC devices may have their
interrupts unmasked. It seems that this interrupt handling is not done
in the kernel for interrupts marked as wake capable, so add support to
store the list of PMIC device interrupts enabled at the time of
entering hibernation and reenable them at the time of exiting
hibernation.Also store polarity of the interrupts at the time of
hibernation and restore back at the time of exit.

Change-Id: I9333b2228aa5b8f5eb4efd00442c2d11ebe3ea35
Signed-off-by: Jishnu Prakash <quic_jprakash@quicinc.com>
Signed-off-by: Ayyagari Ushasreevalli <quic_aushasre@quicinc.com>