msm: mhi_dev: Increase timeout of async transfer wait
On certain targets, there are multiple use cases such as ADB and NBD
on the PF MHI instance. Since there is only one GSI pipe assigned
for PF MHI and there is a lot of traffic, transfer completions
take more time to complete.

A read operation of 37 bytes is initiated by the UCI driver in an
asynchronous manner and waits for a timeout of 100 ms. Since the
transfer does not complete within the 100 ms timeout, the read
operation fails, the channel closes, and pending transfers are marked
as stale.

Since the callback from ECPRI takes around 150 ms, we get a callback
around ~135 ms. The client handle stored in `mreq` is already freed
as part of the channel close, causing a data abort.

Increasing the timeout helps avoid this issue.

Call trace:
[ T4063] Call trace:
[ T4063]  mhi_dev_transfer_completion_cb+0x38/0x7bc
[ T4063]  ecpri_dma_mhi_memcpy_async_wq_cb_ready+0x16c/0x334
[ T4063]  process_one_work+0x1ac/0x3a8
[ T4063]  worker_thread+0x230/0x498
[ T4063]  kthread+0x164/0x228
[ T4063]  ret_from_fork+0x10/0x20.

Change-Id: I4d526b5d11eb4d67e8ef80261d8cab4ad92e7a0a
Signed-off-by: Subramanian Ananthanarayanan <quic_skananth@quicinc.com>