net: stmmac: Set DMA ring size in HW when converting channel for HW path
When a channel moves from HW path to SW path, update the descriptor
ring length in the EMAC HW register. This is a required step because
the ring sizes for HW and SW path can be different and a mismatch
can cause MTL to send packets out of the range of the allocated
descriptor ring.

Change-Id: Ie740e2cd1df5d24baaf51cc4c5b809cab8e3ea43
Signed-off-by: Mohd Ayaan Anwar <quic_mohdayaa@quicinc.com>