net: Add support for 240 MHz 5 GHz channel bring up
When a 240 MHz channel in the 5 GHz band is initiated, the driver
notifies the kernel via cfg80211_ch_switch_notify() using a channel width
of 320 MHz. However, in kernel version 5.15, 320 MHz bandwidth is only
supported for the 6 GHz band. As a result, the kernel's
cfg80211_chandef_usable() function fails this configuration, causing the
VAP to be brought down.

Additionally, the driver advertises a maximum bandwidth of 240 MHz
for the 5 GHz regulatory rule (covering 5490000 KHz to 5730000 KHz). The
kernel API reg_rule_to_chan_bw_flags() prohibits 320 MHz operation
if the maximum bandwidth is below 320 MHz.

To accommodate this special case of 240 MHz operation in the 5 GHz band,
the patch allows 320 MHz operation when the regulatory rule permits
bandwidth greater than 240 MHz. These changes ensure that the kernel
can support 240 MHz channel operation in the 5 GHz band.

Change-Id: Ie1616a613e9e01ed5bdd4bd29da41f4975bd69ed
Signed-off-by: Ashishkumar Vijayakumar <ashivija@qti.qualcomm.com>