interconnect: camsx: Change clock to q-channel control
Currently for the q-channels to be on, all the required
clocks must be enabled. Change the logic so that any/single
clock is enough to power up the q-channels. Also use
proper node data for the enable check, instead of the one
that we were called for.

Change-Id: I26e8d5818f3932d54564a8b2f003b7aaed938569
Signed-off-by: Yavor Trifonov <ytrifono@qti.qualcomm.com>
(cherry picked from commit 0e427a7fb4737d03796def39b59e72c1735098c4)